+++ src/Rule/ValidPool.php
@@ -29,9 +29,9 @@ final class ValidPool
      *
      * @return Valid New rule Validation
      */
-    public function rule(string $field)
+    public function rule(string ...$field)
     {
-        return $this->pool[$field] = new Valid();
+        return $this->set_field_rule(new Valid(), $field);
     }
 
     /**
@@ -41,9 +41,9 @@ final class ValidPool
      *
      * @return Valid New rule Validation
      */
-    public function __invoke(string $field)
+    public function __invoke(string ...$field)
     {
-        return $this->rule($field);
+        return $this->rule(...$field);
     }
 
     /**
@@ -56,5 +56,22 @@ final class ValidPool
     public function __get($name)
     {
         return $this->rule($name);
+    }
+
+    /**
+     * Helper to add multy rule in single method.
+     *
+     * @param Valid              $valid  Instans for new validation rule
+     * @param array<int, string> $fields Fields name
+     *
+     * @return Valid Rule Validation base from param
+     */
+    private function set_field_rule(Valid $valid, array $fields): Valid
+    {
+        foreach ($fields as $field) {
+            $this->pool[$field] = $valid;
+        }
+
+        return $valid;
     }
 }